lesson 5

+++ ms-delivery/src/main/java/ru/geekbrains/spring/market/repositories/PickUpPointRepository.java
@@ -3,5 +3,5 @@ package ru.geekbrains.spring.market.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.geekbrains.spring.market.model.PickUpPoint;
 
-public interface PickUpPointRepository extends JpaRepository<PickUpPoint, Integer> {
+public interface PickUpPointRepository extends JpaRepository<PickUpPoint, Long> {
 }
