+++ ms-delivery/src/main/java/ru/geekbrains/spring/market/repositories/ShopRepository.java
@@ -3,5 +3,5 @@ package ru.geekbrains.spring.market.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.geekbrains.spring.market.model.Shop;
 
-public interface ShopRepository extends JpaRepository<Shop, Integer> {
+public interface ShopRepository extends JpaRepository<Shop, Long> {
 }
